OK, what happened was that when I originally did the S3 stuff, I changed a bunch of stuff but then found a simpler design and so I changed many things back to the way they were.  However, apparently I got something wrong in there.  So instead of trying to track down what I did wrong, I decided to just start from scratch with the file library.  My main work was confined to three (new) files so all I did was restore the other files from the trunk and then modify them again.  You can see from the diffs of this commit that the changes are not that extensive.
All file related tests are working.

NOTE: One of the password policy tests failed (said the generated password didn't have the correct number of something).  When I ran it again, it worked!  It seems like this test should be refactored to a) use a fixed seed for any randomness and b) generate lots of different passwords to make sure they all conform.