Alright.  After having some initial problems trying to get Bouncy Castle to do ECB mode for Blowfish, I've finally figured it out.  I've created a "BlowfishPwsECB class and it gives the same results as the previous BlowfishJ.BlowfishECB.
In BC, the ECB mode is just the normal mode of BlowfishEngine.  The CBC mode is achieved by wrapping a class called CBCBlockCipher around it.

The odd thing is that for the CBC mode, it is necessary to do an endian conversion but for ECB mode it isn't.  I don't know why this should be.  It seems like PasswordSafe itself is some how inconsistent here.  I suppose it is also possible that BC might be.  All I know is...it isn't me! :-)  I've written lots of test cases to demonstrate that the BC stuff is now acting as a drop in replacement for the previous Pws implementations.

Now after a bit of cleanup I'll be on to SHA1.  Which hopefully is much simpler. :-)